Fix Cell physics body creation
This makes it so that cells are correctly set up in physicsJS and hence are drawn correctly.

Issues with drawing came down to incorrect maths due to not taking into account that rectangles are placed using their central position rather than their end position.

The old implementation went down all the branches for each split. The new implementation does one branch at a time, adding all splits for that branch at once. This might not be a good change, I'll be experimenting with changing back now that things are working correctly and I have a stable implementation to fall back on.