Make collision color changes apply correct colors
This is based upon defensive/offensive colors.
Offensive: Red, White, Grey
Defensive: Blue

If an offensive branch hits a cell, it will temporarily change it's entire cell color to match that branch.
If a defensive branch is hit by an offensive branch, it will temporarily change it's entire cell color to match the defensive branch.
If a cell is hit by an offensive branch, but not on a defensive branch, it will temporarily change it's entire cell color yellow.

These changes will indicate to the user which things are being hit so they understand what is happening.